And update the ones that were not caught by the rename action
